Finish ripping out the old timing
current time is now stored in the app state, and is updated by set interval.

this means our 'time-since' calcs can be pure (yay)

pull all the mutations and reads into a model ns because getters and
setters aren't the end of the world if you have non-trivial state
changes I suppose.

closes #4